requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-documentai==2.31.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
idna==3.7
natsort==8.4.0
numpy==2.0.1
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.2
